+++ src/lib/services/username/lcd.ts
@@ -1,23 +1,28 @@
-import { zBechAddr } from "lib/types";
-import type { Addr } from "lib/types";
+import { getMoveViewJsonLcd } from "../move/module/lcd";
+import type { HexAddr, HexAddr32 } from "lib/types";
+import { zHexAddr } from "lib/types";
 import { parseWithError } from "lib/utils";
 
-const USERNAME_MODULE =
-  "0x42cd8467b1c86e59bf319e5664a09b6b5840bb3fac64f5ce690b5041c530565a";
 const INITIATION_1_LCD = "https://lcd.initiation-1.initia.xyz";
+const USERNAME_MODULE_ADDRESS =
+  "0x42cd8467b1c86e59bf319e5664a09b6b5840bb3fac64f5ce690b5041c530565a" as HexAddr32;
 
-export const getInitiaUsernameByAddress = async (address: Addr) => {
-  const { createInitiaUsernamesClient } = await import("@initia/utils");
+export const getInitiaUsernameByAddress = async (address: HexAddr) =>
+  getMoveViewJsonLcd(
+    INITIATION_1_LCD,
+    USERNAME_MODULE_ADDRESS,
+    "usernames",
+    "get_name_from_address",
+    [],
+    [`"${address}"`]
+  ).then((data) => (data ? `${data}.init` : null));
 
-  const client = createInitiaUsernamesClient(INITIATION_1_LCD, USERNAME_MODULE);
-  return client.getUsername(address);
-};
-
-export const getAddressByInitiaUsername = async (username: string) => {
-  const { createInitiaUsernamesClient } = await import("@initia/utils");
-
-  const client = createInitiaUsernamesClient(INITIATION_1_LCD, USERNAME_MODULE);
-  return client
-    .getAddress(username)
-    .then((data) => (data ? parseWithError(zBechAddr, data) : null));
-};
+export const getAddressByInitiaUsername = async (username: string) =>
+  getMoveViewJsonLcd(
+    INITIATION_1_LCD,
+    USERNAME_MODULE_ADDRESS,
+    "usernames",
+    "get_address_from_name",
+    [],
+    [`"${username.toLowerCase().replace(".init", "")}"`]
+  ).then((data) => (data ? parseWithError(zHexAddr, data) : null));